+++ api/app/settings/common.py
@@ -304,7 +304,7 @@ TASK_PROCESSOR_DATABASES = env.list(
 
 
 LOGIN_THROTTLE_RATE = env("LOGIN_THROTTLE_RATE", "20/min")
-DCR_THROTTLE_RATE = env("DCR_THROTTLE_RATE", "10/min")
+DCR_THROTTLE_RATE = env("DCR_THROTTLE_RATE", "500/month")
 SIGNUP_THROTTLE_RATE = env("SIGNUP_THROTTLE_RATE", "10000/min")
 USER_THROTTLE_RATE = env("USER_THROTTLE_RATE", default=None)
 MASTER_API_KEY_THROTTLE_RATE = env("MASTER_API_KEY_THROTTLE_RATE", default=None)
